requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-use-email-as-username==1.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-excel==2.4.0
drf-renderer-xlsx==1.0.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.3
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
XlsxWriter==3.2.0
xlwt==1.3.0
